+++ src/activitypub/views.rs
@@ -178,6 +178,7 @@ async fn outbox(
             collection_id,
             Some(first_page_id),
             None,
+            false,
         );
         let response = HttpResponse::Ok()
             .content_type(AP_MEDIA_TYPE)
@@ -249,6 +250,7 @@ async fn followers_collection(
         collection_id,
         None,
         Some(user.profile.follower_count),
+        false,
     );
     let response = HttpResponse::Ok()
         .content_type(AP_MEDIA_TYPE)
@@ -277,6 +279,7 @@ async fn following_collection(
         collection_id,
         None,
         Some(user.profile.following_count),
+        false,
     );
     let response = HttpResponse::Ok()
         .content_type(AP_MEDIA_TYPE)
@@ -305,6 +308,7 @@ async fn subscribers_collection(
         collection_id,
         None,
         Some(user.profile.subscriber_count),
+        false,
     );
     let response = HttpResponse::Ok()
         .content_type(AP_MEDIA_TYPE)
@@ -329,6 +333,7 @@ async fn featured_collection(
             collection_id,
             Some(first_page_id),
             None,
+            true,
         );
         let response = HttpResponse::Ok()
             .content_type(AP_MEDIA_TYPE)